// front/src/app/secure/applicant/application/intake-applications/attachment.interface.ts
import {AttachmentType} from './attachment-type.enum';
import {MetaObject} from '../../../../core/meta-object.interface';
export interface Attachment extends MetaObject {
  id: number;
  name: string;
  mimeType: string;
  size: number;
  attachmentType: AttachmentType;

  // transient
  selected?: boolean;
}